% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_strata_data}
\alias{create_strata_data}
\title{Create Strata Data}
\usage{
create_strata_data(n_axes, grouping_levels)
}
\arguments{
\item{grouping_levels}{A vector of values for the grouping variable.}

\item{n_strata}{The number of strata to create.}

\item{group_var}{The variable to use for grouping the strata.}
}
\value{
A data frame with strata positions and grouping variable.
}
\description{
This function generates a simple data frame for strata, useful for testing or quick visualization.
}
