% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_alluvial_from_data}
\alias{plot_alluvial_from_data}
\title{Plot Alluvial Diagram from Data with User-Defined Strata Order}
\usage{
plot_alluvial_from_data(
  input_data,
  omics_order,
  strata_order = NULL,
  title = NULL
)
}
\arguments{
\item{input_data}{A data frame containing flow information.}

\item{omics_order}{A character vector specifying the order of omics layers.}

\item{strata_order}{A named list where names are omics layers and values are character vectors
specifying the desired order of strata within each layer.}
}
\value{
A ggplot object representing the alluvial plot.
}
\description{
This function takes flow data and orders for omics layers and strata to
generate an alluvial plot.
}
